requirements.txt
cs50
flask 
flask_session
werkzeug.security
functools
